Fix leaky abstraction in interaction factory
Previously the factory depended on the component having a reference to a
DOM element named "canvas", which breaks the encapsulation of the
component. Now the canvas DOM element must be passed in as an argument.

This moves handler creation to after mounting, rather than at
construction, and necessitates a forceUpdate.