Refactor interactive canvas to component
Previously adding interactivity via the canvas was a higher order
component. However that led to encapsulation problems when we want to
render more than just the basic canvas and points (i.e. for getting a
point on a curve, we need a t parameter).

The state was held by the higher order component, but now the
InteractiveCanvas is a thin layer without state. Instead, the callbacks
are used to decouple concerns. It's a management layer for the canvas
DOM element, and is responsible for setting up the mouse event handlers.

The user now needs to compose the InteractiveCanvas, passing in a
callback for getting points, handling when a point moves, and handling
drawing the canvas.